avoid default http port 80; fix bugs that caused lost/misconfigured mode and .conf files

+++ libs/configuration_app/app.py
@@ -4,6 +4,7 @@ import os
 import time
 from threading import Thread
 import fileinput
+import syslog
 
 app = Flask(__name__)
 app.debug = True
@@ -105,12 +106,17 @@ def create_wpa_supplicant(ssid, wifi_key):
     os.system('mv wpa_supplicant.conf.tmp /etc/wpa_supplicant/wpa_supplicant.conf')
 
 def set_ap_client_mode():
-    os.system('rm -f /etc/raspiwifi/host_mode')
-    os.system('rm /etc/cron.raspiwifi/aphost_bootstrapper')
-    os.system('cp /usr/lib/raspiwifi/reset_device/static_files/apclient_bootstrapper /etc/cron.raspiwifi/')
-    os.system('chmod +x /etc/cron.raspiwifi/apclient_bootstrapper')
-    os.system('mv /etc/dnsmasq.conf.original /etc/dnsmasq.conf')
-    os.system('mv /etc/dhcpcd.conf.original /etc/dhcpcd.conf')
+    if os.path.exists("/etc/raspiwifi/host_mode"):
+        os.system('rm -f /etc/raspiwifi/host_mode')
+        os.system('rm -f /etc/cron.raspiwifi/aphost_bootstrapper')
+        os.system('cp /usr/lib/raspiwifi/reset_device/static_files/apclient_bootstrapper /etc/cron.raspiwifi/')
+        os.system('chmod +x /etc/cron.raspiwifi/apclient_bootstrapper')
+        os.system('rm -f /etc/dnsmasq.conf')
+        os.system('mv /etc/dnsmasq.conf.original /etc/dnsmasq.conf')
+        os.system('rm -f /etc/dhcpcd.conf')
+        os.system('mv /etc/dhcpcd.conf.original /etc/dhcpcd.conf')
+        syslog.syslog('raspiwifi - set_ap_client_mode')
+
     os.system('reboot')
 
 def update_wpa(wpa_enabled, wpa_key):
